bugfix: distorted paths when fetching from firebase

+++ app/src/main/java/dev/borisochieng/sketchpad/collab/data/models/DBOffset.kt
@@ -1,6 +1,6 @@
 package dev.borisochieng.sketchpad.collab.data.models
 
 data class DBOffset(
-    val x: Double = 0.0,
-    val y: Double = 0.0
+    val x: Float = 0f,
+    val y: Float = 0f
 )
